+++ src/core/domain/models/User/User.cs
@@ -1,7 +1,7 @@
-using domain.models.Users.values;
+using domain.models.user.values;
 using OperationResult;
 
-namespace domain.models.Users;
+namespace domain.models.user;
 
 public class User
 {
