+++ src/core/domain/models/WorkItem/WorkItem.cs
@@ -1,4 +1,4 @@
-using domain.models.Users;
+using domain.models.user;
 using domain.models.workItem.values;
