+++ src/tests/Unit/Unit.csproj
@@ -29,6 +29,7 @@
 
     <ItemGroup>
       <Folder Include="models\WorkItem\" />
+      <Folder Include="values\WorkSpace\" />
     </ItemGroup>
 
 </Project>
