+++ src/tests/Unit/values/User/EmailTests.cs
@@ -1,5 +1,5 @@
 using domain.exceptions.User.Email;
-using domain.models.Users.values;
+using domain.models.user.values;
 
 namespace Unit.values.User;
