+++ src/tests/Unit/values/User/FirstNameTests.cs
@@ -1,5 +1,5 @@
 using domain.exceptions.User.FirstName;
-using domain.models.Users.values;
+using domain.models.user.values;
 
 namespace Unit.values.User;
